requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
appium-python-client==1.0.2  # via -r requirements.in
click==7.1.2              # via pip-tools
pip-tools==5.4.0          # via -r requirements.in
selenium==3.141.0         # via appium-python-client
six==1.15.0               # via pip-tools
urllib3==1.26.2           # via selenium

# The following packages are considered to be unsafe in a requirements file:
# pip
